Fixed a bug in cgi.c ( or, possibly, worked around a bug in various browsers by changing cgi.c -- your call ):  we can't just test for "cookie == NULL" to see if the user's cookie has been clobbered ( by logout, presumably ) because iCab and OmniWeb, at least, are broken and ignore the Expires header.  *shrug*
Changed the Error template so it doesn't say 'Error' but uses the value for title in both title positions.  This lets us use it as a splash screen, as a login succeeded screen, etc.

Added a link to login from the splash screen.

Added a link back to the desired resource from the 'login succeeded'
screen.